Refactor home screen and add library with songs template
